exploration/requirement.txt
absl-py==0.7.1
alabaster==0.7.12
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.2.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
conda==4.3.16
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
flatbuffers==1.11
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
grpcio==1.21.1
gym==0.13.0
h5py==2.9.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
joblib==0.13.2
jsonpickle==0.9.6
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.0.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==7.0.0
munch==2.3.2
mysql-connector-python==8.0.16
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.4
numpydoc==0.9.1
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.2
ptyprocess==0.6.0
py==1.8.0
py-cpuinfo==5.0.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
PyQt5==5.11.2
PyQt5-sip==4.19.17
PyQtWebEngine==5.12.1
pyrsistent==0.15.2
pytest==4.6.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
pyzmq==18.0.1
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
ray==0.7.1
redis==3.2.1
requests==2.22.0
retrying==1.3.3
rope==0.14.0
ruamel.yaml==0.15.96
sacred==0.7.4
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==2.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.4
spyder-kernels==0.4.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.5
typing==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
zipp==0.5.1
